club-activity-backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
PyJWT==2.8.0
python-dotenv==1.0.1
Werkzeug==2.3.7